add persisted_entries
